+++ app/cells/button/submit_cell.rb
@@ -1,0 +1,13 @@
+module Button
+  class SubmitCell < BaseCell
+    private
+
+    def icon_name
+      'check-circle'
+    end
+
+    def form
+      model
+    end
+  end
+end