+++ app/cells/concerns/cell/font_awesome.rb
@@ -1,0 +1,5 @@
+module Cell::FontAwesome
+  private
+
+  define_method :icon, FontAwesome::Sass::Rails::ViewHelpers.instance_method(:icon)
+end