+++ app/cells/icon_label_cell.rb
@@ -1,0 +1,13 @@
+class IconLabelCell < Cell::ViewModel
+  include Cell::FontAwesome
+
+  private
+
+  def icon_name
+    options[:icon]
+  end
+
+  def text
+    model
+  end
+end